fix: make exports work as documented
According to the readme, usage was supposed to work like this:

```js
import parse from 'style-to-js';

parse('line-height: 42');
```

But in reality, it worked like this:

```js
import parse from 'style-to-js';

parse.default('line-height: 42');
```

These changes make usage work as documented. This also makes usage nicer
from CJS.

The `default` property was added for backwards compatibiliy, along with
a test.

`@rollup/plugin-typescript` can’t handle TypeScript’s `export =` syntax.
As a workaround, the CJS output is now used as the input for the UMD
bundle.